requirements.txt
asgiref==3.2.7
coverage==5.1
Django==3.0.6
djangorestframework==3.11.0
Pillow==7.2.0
pkg-resources==0.0.0
pytz==2020.1
sqlparse==0.3.1
